article_src/requirements.txt
# Python 3.11
python-dotenv
transformers
datasets
torch
accelerate
tokenizers
evaluate
bitsandbytes
evaluate
scikit-learn
python-dotenv